Add files via upload
Command line rectangle "drawing" program. Employs the use of objects and prints out rectangles of pre-defined sizes as created within the code, as well as an interactive feature for the user to create a rectangle of dimensions that they choose and characters for the horizontal & vertical lines, along with the corner characters.