Add files via upload
Command line dice histogram. The user enters the number of "rolls", and the program calculates the percentage and number of rolls certain values are rolled for two dice given the user's input. Also includes a primitive bar chart that visually displays the the percentage of a rolled number for that instance of the dice rolling game.